Moved channel setups to their own file. Moved channel colors to their own file so they can be shared. Added some more explicit type specifiers to replace old JSDocs
